+++ package.json
@@ -4,7 +4,7 @@
   "type": "module",
   "main": "index.js",
   "scripts": {
-    "dev": "ts-node-dev --respawn src/app.ts",
+    "build": "tsc",
     "start": "node --enable-source-maps dist/app.js"
   },
   "repository": {
@@ -20,7 +20,7 @@
   "homepage": "https://github.com/YSJ0228/6-sprint-mission-be#readme",
   "description": "",
   "dependencies": {
-    "@prisma/client": "^6.7.0",
+    "@prisma/client": "^6.10.1",
     "@types/bcrypt": "^5.0.2",
     "@types/body-parser": "^1.19.6",
     "@types/cookie-parser": "^1.4.9",
@@ -40,6 +40,7 @@
   "devDependencies": {
     "@types/express": "^5.0.3",
     "prisma": "^6.7.0",
+    "ts-node": "^10.9.2",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.8.3"
   },
