+++ src/app.ts
@@ -1,0 +1,36 @@
+import express, { Express } from "express";
+import cors from "cors";
+import cookieParser from "cookie-parser";
+import path from "path";
+import { fileURLToPath } from "url";
+import routes from "./routes/index.js";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+process.on('uncaughtException', (err) => {
+  console.error('Uncaught Exception:', err);
+});
+process.on('unhandledRejection', (reason, promise) => {
+  console.error('Unhandled Rejection:', reason);
+});
+
+const app: Express = express();
+const PORT: number = 5000;
+
+app.use(
+  cors({
+    origin: "http://localhost:3000",
+    credentials: true,
+  })
+);
+app.use(express.json());
+app.use(cookieParser());
+
+// 'uploads' 디렉토리를 정적 파일 경로로 설정합니다.
+app.use("/uploads", express.static(path.join(__dirname, "../uploads")));
+app.use("/api", routes);
+
+app.listen(PORT, () => {
+  console.log(`✅ 서버가 http://localhost:${PORT} 에서 실행 중입니다.`);
+});
