+++ src/middlewares/auth.middleware.ts
@@ -1,0 +1,32 @@
+import { Request, Response, NextFunction } from 'express';
+import jwt, { JwtPayload } from 'jsonwebtoken';
+import { User } from '@prisma/client';
+
+const SECRET_KEY = process.env.JWT_SECRET || "panda-secret";
+
+// Express Request 타입 확장
+declare global {
+  namespace Express {
+    interface Request {
+      user?: User | JwtPayload;
+    }
+  }
+}
+
+export function verifyToken(req: Request, res: Response, next: NextFunction) {
+  const authHeader = req.headers.authorization;
+
+  if (!authHeader || !authHeader.startsWith("Bearer ")) {
+    return res.status(401).json({ message: "로그인이 필요합니다." });
+  }
+
+  const token = authHeader.split(" ")[1];
+
+  try {
+    const decoded = jwt.verify(token, SECRET_KEY);
+    req.user = decoded as User;
+    next();
+  } catch (error) {
+    return res.status(401).json({ message: "유효하지 않은 토큰입니다." });
+  }
+}
