--- src/routes/auth.route.js
@@ -1,9 +1,0 @@
-import express from "express";
-import { signup, signin } from "../controllers/auth.controller.js";
-
-const router = express.Router();
-
-router.post("/signup", signup);
-router.post("/signin", signin);
-
-export default router;
